Add decorator features.
Now Santa’s built-in logger type can create a decorator that wraps a copy of itself, with very low overhead. Using the decorator feature, you can even create a decorator (the shadow of the logger) for each transaction processed by the application, and then change some of the logger parameters (for example: the log level of the logger, etc.) to achieve the logger parameters Fine-grained control of the runtime. For details, please refer to the comment section of the `Decorator` function and `Decorator` structure provided by the logger.